@@ -6,15 +6,23 @@
 
     <groupId>toutouchien.niveriaapi</groupId>
     <artifactId>NiveriaAPI</artifactId>
-    <version>1.10.0</version>
+    <version>2.0.0</version>
     <packaging>jar</packaging>
 
     <name>NiveriaAPI</name>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+
+    <pluginRepositories>
+        <pluginRepository>
+            <id>bytecode.space</id>
+            <url>https://repo.bytecode.space/repository/maven-public/</url>
+        </pluginRepository>
+    </pluginRepositories>
+
     <build>
         <plugins>
             <plugin>
