@@ -1,5 +1,6 @@
 package toutouchien.niveriaapi.delay;
 
+import net.kyori.adventure.text.Component;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
@@ -8,6 +9,7 @@ import java.util.function.Consumer;
 
 public class Delay {
 	private final Player player;
+	private final Component text;
 	private final Consumer<Player> successConsumer;
 	private final Consumer<Player> failConsumer;
 	private final Location originalLocation;
@@ -17,8 +19,9 @@ public class Delay {
 	private final int delay;
 	private int delayRemaining;
 
-	Delay(@NotNull Player player, Consumer<Player> successConsumer, Consumer<Player> failConsumer, int delay, boolean cancelOnMove, boolean actionbar, boolean chat, boolean title) {
+	Delay(@NotNull Player player, Component text, Consumer<Player> successConsumer, Consumer<Player> failConsumer, int delay, boolean cancelOnMove, boolean actionbar, boolean chat, boolean title) {
 		this.player = player;
+		this.text = text;
 		this.successConsumer = successConsumer;
 		this.failConsumer = failConsumer;
 		this.originalLocation = player.getLocation();
@@ -38,6 +41,10 @@ public class Delay {
 
 	public Player player() {
 		return player;
+	}
+
+	public Component text() {
+		return text;
 	}
 
 	public Consumer<Player> successConsumer() {
